feat: make shouldDoInterceptionBasedOnUrl overrideable
