+++ lib/ts/utils/index.ts
@@ -24,7 +24,7 @@ import {
     RecipePreAPIHookFunction
 } from "../types";
 import WindowHandlerReference from "../utils/windowHandler";
-import { enableLogging, logDebugMessage } from "../logger";
+import { enableLogging } from "../logger";
 
 export function normaliseURLDomainOrThrowError(input: string): string {
     let str = new NormalisedURLDomain(input).getAsStringDangerous();
@@ -167,58 +167,6 @@ export function validateAndNormaliseInputOrThrowError(options: InputType): Norma
         onHandleEvent,
         override
     };
-}
-
-export function shouldDoInterceptionBasedOnUrl(
-    toCheckUrl: string,
-    apiDomain: string,
-    sessionTokenBackendDomain: string | undefined
-): boolean {
-    logDebugMessage(
-        "shouldDoInterceptionBasedOnUrl: toCheckUrl: " +
-            toCheckUrl +
-            " apiDomain: " +
-            apiDomain +
-            " sessionTokenBackendDomain: " +
-            sessionTokenBackendDomain
-    );
-    function isNumeric(str: any) {
-        if (typeof str != "string") return false; // we only process strings!
-        return (
-            !isNaN(str as any) && !isNaN(parseFloat(str)) // use type coercion to parse the _entirety_ of the string (`parseFloat` alone does not do this)...
-        ); // ...and ensure strings of whitespace fail
-    }
-
-    // The safest/best way to add this is the hash as the browser strips it before sending
-    // but we don't have a reason to limit checking to that part.
-    if (toCheckUrl.includes("superTokensDoNotDoInterception")) {
-        return false;
-    }
-
-    toCheckUrl = normaliseURLDomainOrThrowError(toCheckUrl);
-    let urlObj = new URL(toCheckUrl);
-    let domain = urlObj.hostname;
-    if (sessionTokenBackendDomain === undefined) {
-        domain = urlObj.port === "" ? domain : domain + ":" + urlObj.port;
-        apiDomain = normaliseURLDomainOrThrowError(apiDomain);
-        let apiUrlObj = new URL(apiDomain);
-        return domain === (apiUrlObj.port === "" ? apiUrlObj.hostname : apiUrlObj.hostname + ":" + apiUrlObj.port);
-    } else {
-        let normalisedsessionDomain = normaliseSessionScopeOrThrowError(sessionTokenBackendDomain);
-        if (sessionTokenBackendDomain.split(":").length > 1) {
-            // means port may provided
-            let portStr = sessionTokenBackendDomain.split(":")[sessionTokenBackendDomain.split(":").length - 1];
-            if (isNumeric(portStr)) {
-                normalisedsessionDomain += ":" + portStr;
-                domain = urlObj.port === "" ? domain : domain + ":" + urlObj.port;
-            }
-        }
-        if (sessionTokenBackendDomain.startsWith(".")) {
-            return ("." + domain).endsWith(normalisedsessionDomain);
-        } else {
-            return domain === normalisedsessionDomain;
-        }
-    }
 }
 
 export function getNormalisedUserContext(userContext?: any): any {
